refactor AbstractRPcRemotingClient.
